+++ src/options/rome.ts
@@ -1,3 +1,5 @@
-export default interface ROME {
+import type { Configuration } from "@rometools/wasm-nodejs";
+
+export default interface ROME extends Configuration {
 	[key: string]: any;
 }